+++ src/components/BulkEditProfiles/BulkEditProfilesPane.js
@@ -91,6 +91,7 @@ export const BulkEditProfilesPane = ({
     isFetching,
     isLoading: isProfilesLoading,
     profiles,
+    refetch,
   } = useBulkEditProfiles({ entityType });
 
   const userIds = useMemo(() => profiles.map(profile => profile.updatedBy), [profiles]);
@@ -240,6 +241,7 @@ export const BulkEditProfilesPane = ({
             <BulkEditProfileDetails
               entityType={entityType}
               onClose={onCloseDetailsPane}
+              refetch={refetch}
             />
           </Layer>
         )}