+++ src/hooks/api/index.js
@@ -1,5 +1,6 @@
 export * from './useBulkEditLogs';
 export * from './useBulkEditProfile';
+export * from './useBulkEditProfileMutation';
 export * from './useBulkEditProfiles';
 export * from './useBulkOperationDetails';
 export * from './useBulkOperationStart';